reservation-service/requirements.txt
alembic==1.14.0
aniso8601==9.0.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.1
bidict==0.23.1
black==24.10.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
constantly==23.10.4
coverage==7.6.8
cryptography==43.0.3
daphne==4.1.2
dnspython==2.7.0
eventlet==0.38.0
factory_boy==3.3.1
Faker==33.0.0
flake8==7.1.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
importlib_resources==6.4.5
incremental==24.7.2
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.0
pyOpenSSL==24.2.1
pytest==8.3.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
referencing==0.35.1
rpds-py==0.21.0
service-identity==24.2.0
setuptools==75.6.0
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.36
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
zope.interface==7.1.1
